chore: refactor locust stress test
